feat: environment and modules
